Rename Env⟦ Γ ⟧ to ⟦ Γ ⟧Context (see #18).
The meaning of a context is the type of environments that provide
bindings for all variables declared in the context. The old name
focused on the fact that the meaning of a context is an environment.
The new name focuses on the fact that we are talking about the
meaning of contexts.

Old-commit-hash: c2044ece8a26fe1fab4b62eed31cb4a3bb660fd3