Prove and use weakening (fix #22 and #29).
The key ideas for this implementation are to (1) reify weakening
operations as a subcontext relationship _≼_ and to (2) generalize
the type of Δ to allow weakening to happen before Δ is
invoked.

Details on (1): Γ₁ ≼ Γ₂ specifies how the elements common to Γ₁
and Γ₂ are interleaved into the elements of Γ₂.

Details on (2): The type of Δ is generalized to:

  Δ : ∀ {Γ₁ Γ₂ τ} →
    {{Γ′ : Δ-Context Γ₁ ≼ Γ₂}} →
    (t : Term Γ₁ τ) →
    Term Γ₂ (Δ-Type τ)

This allows us to merge a weakening operation that meets a Δ with
the weakening operation inside the Δ, using ≼-trans. The
implementation of weakening for Δ terms is:

  weaken Γ′ (Δ {{Γ″}} t) = Δ {{≼-trans Γ″ Γ′}} t

Old-commit-hash: 64f3fc83db14b79d41938d924b176058ef86f747