agda: show that diff-is-valid is the right thing to prove (#33)
derive-is-valid needs to be proved using induction, but as induction
hypothesis it's not strong enough and a hole is left in the proof.

However, another conjecture is the validity of the result of diff,
diff-is-valid: it is a slight generalization of derive-is-valid, and
as shown in this commit, would be sufficient as induction hypothesis.

Hence, in subsequent commits I will try to prove diff-is-valid
instead (and I will succeed).

Old-commit-hash: 470e7716f46766a6ca912382551d466ed5a8068e