agda: prove variant of diff-apply on weakly valid changes (#33)
This is a major step forward for diff-apply. The annoyance is that the
new equality is probably not a congruence, but this should not be a
real problem with a different proof structure.

See discussion at:
https://github.com/ps-mr/ilc/commit/25b473eeff3f5286dafba706c0d7c64b886d29b3#commitcomment-3291862

Old-commit-hash: 5ebd7c2a74db46b3e6e2d5c373f4c5f94a1cb935