Standalone semantic domain for folds on bags (with derivative) WIP
Develop bags in the semantic domain, independently of the calculus.

* Try develop a reasonably simple semantics for bags: a pair of sorted
  lists, one with present elements (each entry has multeplicity +1)
  and one with absent elements (each entry has multeplicity -1).

* Define monoid homomorphism (into a group), `hom`, and its
  derivative, `homDelta`. The definition turns out to be really easy,
  and the derivative is the same function if changes are just bags.

* State the correctness theorem for `homDelta`.

* Define a few functions (like map) in terms of `hom`.

  Note that however we might still need a few other primitives.

Some aborted attempts are still in, in case we prefer them in the end;
the list using multiplicities might be relevant here.

Note: This just considers the derivative wrt. changes in the bag, not
wrt. changes in the function. AFAIK, the case where the function
changes can't be optimized, hence there we can simply write the
obviously correct definition for the derivative.

Old-commit-hash: b9249eb576720c2f0b4e97a04d50106cd9c3fbe1