Example 3 works noŵ¹. ̂¹ Other than that the `Bag` type is an abstract postulate.
A presentation of example 3 starts from line 297.

The optimized derivation `derive2` considers the derivative of
a term to be the nil change only if it is closed. Should we have
β-equivalence test or some inherited description of whether the
argument to an abs-term may change, we could handle more examples.

Old-commit-hash: 6488c110b93dc4c760ddc61feef198b55ab4940d