Use a family of constants instead of type-of.
This design is more regular, because it uses the same mechanism for
constants, variables and terms. Before this changes, constants used
type-of while variables and terms had been indexed by object types.

I guess that this design is also easier to use, because it avoids
unification involving type-of.

The set of constants of a calculus is now of type (Type B -> Set)
instead of Set.

Old-commit-hash: e008c64abf614557689caaf464bc357a69399faa