Move all of S.C.Plotkin into P.S.Type.
If a calculus has types, it automatically has typing contexts. No need
to put this in a different module.

Syntax.Context.Plotkin only contained a single function anyway, and that
function deals with converting context structure to type structure, so
it can equally well live in the module that defines the type structure
instead of the module that defines the context structure.

Old-commit-hash: 87993dc07f411d0bf0d264bab7ec02a7299eebc2