Remove script generating README.agda
By now we decided that README.agda should be edited by hand, so it
shouldn't be generated.
Hence the extra logic in this script is not needed any more.
Now we switch away from Ruby, which has the nice effect of reducing
dependencies - we depend on Haskell instead, which is required for Agda
anyway.

Old-commit-hash: 86e551308995c5c9eec8f068698e0c62d93108cf