Provide two lists of terms to deriveConst.
Before this change: deriveConst accepts a list of terms that contains
two entries for every subterm of the constant, one with the subterm, one
with the derivative of the subterm.

After this change: deriveConst accepts two lists of terms, one with all
the subterms, and one with all the derivatives.

The new structure fits better with ⟦_⟧ΔConst.

Old-commit-hash: 3682bbf0ae0a0728bc41a393bc85e575c79a0b0a