Agda: start reflecting back changes to the main theorem in the paper
We simplified the statement in the paper by talking about valid changes.
Though this is a minor change, this should be reflected in Agda, also to
get an exact statement for the paper.

This commit discusses what to do; I didn't try writing the real
statement.

Old-commit-hash: 0d78ce6f1fab80aa61430ee3241ca2c7080028ba