Pick correct base case for recursive functions
A little duplication remained, because absN and absV do not have as base
case arity 0 but arity 1, just because the case for arity 0 is not very
useful per se. Hence, abs{N,V} need to unroll one recursive invocation
into the base case, and need to have an helper (absNBase) to factor out
the ensuing duplication.

This commit fixes that; next commit will inline absNBase.