Add extension points for nil primitive
This commit adds nil to the basic operations of the plugin interface,
like apply and diff.

This implies adding various other related components, and adding
implementations for Nehemiah.

This operation is not yet used anywhere - in particular, the algebra of
changes still has a fixed definition, so proofs need not be
altered *yet*.

More extension points will be added later.