Add script to simply regen Everything.agda
After calling this you can call agda on README.agda separately --- this
is useful when agda is open inside Emacs, since rechecking is then much
faster. My previous workflow for this case was a hack: either start
agdaCheck.sh and kill it after regeneration, or forget to call it and
check with a stale Everything.agda.

This doesn't belong in the branch, but I'm ignoring this minor issue, at
least for now.