Travis: Check if the included happy/alex are good
Goal: avoid rebuilding them at each test (they don't contain libraries,
so cabal does not figure out they don't need be reinstalled).

Unfortunately, I should have done this earlier, because now we have
built versions in the cache (see the hack calling the global versions
directly). That's a problem because if only the versions in the cache
are good, clearing the cache would break the build.