Simplify defining derivatives for constants
At this point defining derivatives for constants is quite annoying. We
define by hand what's the semantics for derivatives of constants, and
prove by hand lemmas to show those derivatives are indeed nil changes.
Unary functions take a lemma to show they're correct function changes
for their first argument. Binary functions take a second lemma to show
they're correct in their second argument. Finally, we must show they're
nil.

Instead, just state generally that the differential semantics of each
constant `c` is its nil change (and prove once and for all this is a
correct change semantics) and specify the action of `derive(c)`. The nil
change provides, once and for all, a correct semantics and proofs that
it's a valid change.

This way, the only lemma we have to prove only relates the differential
semantics of each constant with the semantics of `derive(c)` via the
existing erasure relation. That is, we simply prove that the derivative
we define is indeed a derivative, mostly without additional overheads
due to proofs embedded in functions. The only exception is the proof for
flatmap: there the proof that argument functions are related does *not*
say that the arguments are literally the same, so we have a bit more
overhead, that is however essential since we must use a logical relation
here.