Remove some level polymorphism causing ambiguity
We currently allow the type of `Δ x` to be at a different level than the
type of `x`. This flexibility appears not needed (though you need to
drop it in just the right way). Moreover, this flexibility means that
instance parameter can't be solved by instance inference (since the
level for `Δ x` is still unsolved when we search for the instance, and
Agda since a while doesn't solve instance arguments for types with
unsolved metas).

Drop this, and remove annotations that are now redundant.