Simplify proofs that realizers are changes
To define derivatives of primitives, we want to provide the right
realizers (that is the derivatives, without embedded proofs) and prove
they are correct.

Instead of having to prove by hand that realizers are correct changes
for each of their argument, and then that they're derivatives, it's
easier to just prove that they give results that are change-equivalent
to the derivative's results.

This commit adds lemmas showing that a realizer with such a proof gives
a correct change.