Add simpler proof
By separating validity proofs from changes, and making validity a
logical relation (though this is not immediately visible), give a
simpler proof of correctness of differentiation.

== Stats ==

Actual time taken for this formalization: 1 day.
Line count: 428 LoC (including blanks and comments).

Both counts exclude the code reused by inclusion. Much more code was
reused by copy-pasting.

== Highlights ==

This proof is carefully set up so that (with made-up notation) the
following equation:

⟦ Δ τ ⟧ = Δ ⟦ τ ⟧

holds *definitionally* in Agda. This way we don't need to have both
derive and a separate change semantics, and we can define a validity
relation relating ⟦ τ ⟧ and ⟦ Δ τ ⟧ directly.

I arrived at this proof by studying the original one and trying to
simplify its presentation for my PhD thesis. The last push for this
proof came from a 6-page pen-and-paper proof circulated by Yufei Cai
some time ago; however, that proof was purely set-theoretic, and none of
us had realized that this could be formalized so easily, or that it
contained a logical relation. (However, Cai had written earlier a few
much smaller proofs, though they seemed to have significant
limitations).

== Downsides ==

Right now, this proof doesn't define syntactic change structures, so I
can't write, as in the paper:

⟦ f a ⊕ Derive(f) a da ⟧ ≡ ⟦ f (a ⊕ da) ⟧

but only

⟦ f a ⟧ ⊕ ⟦ Derive(f) a da ⟧ ≡ ⟦ f ⟧ (⟦ a ⟧ ⊕ ⟦ da ⟧)

At the time we cared a lot for proving the former equation. Doing that
last step would add some annoyance, but at least we'd only need one
object-level copy of ⊕, and we wouldn't need to define its semantics ⟦⊕⟧
as an erased ⊕. Nevertheless, to relate object-level ⊕ and meta-level ⊕
we'd probably need to copy one of the existing inductive proofs.

I also simplified quite a few other things:

* no constants are given yet
* avoid the plugin structure
* avoid defining families of change structures (this was clever when it
  worked, but it totally stopped working ages ago); this simplifies
* sometimes, use less Unicode for names than in the other proof
* some names are probably still bad.