Even simpler correctness proof
Sums are not in. Still, even for the same language, the difference is
amazing. `correctDerive` + `validDerive` take 70 lines, while
`fromtoDerive` takes around 10. And this file includes a definition of
validity (though not of change types)!

Making change environments extend the base environments (as Cai, Yann
and finally I did on paper) might even simplify auxiliary lemmas a bit
more.

Of course, as we know, this doesn't even define ⊕, let alone prove its
properties, recover the lemmas we'd want about derivatives, and so on.