WIP implement composition for sum changes STUCK
Goal: implement change composition without using ⊝ for any datatype, so
that we can also drop the base value index for compose.

Problem: if I compose a change from inj₁ a1 to inj₂ b2, and a change
from inj₂ b2 to inj₁ a3, I need to produce a change from inj₁ a1 to inj₁
a3. I can either use a3 ⊝ a1, and we don't want, or produce a
replacement change from a inj₁ to inj₁, which is not valid. So I'll need
to widen validity.