Conclude that derive is correct from fundamental property
For a description I'll just paste the head comment for the new module.

In this module we conclude our proof: we have shown that t, derive-dterm t and
produce related values (in the sense of the fundamental property,
rfundamental3), and for related values v1, dv and v2 we have v1 ⊕ dv ≡
v2 (because ⊕ agrees with validity, rrelV3→⊕).

We now put these facts together via derive-correct-si and derive-correct.
This is immediate, even though I spend so much code on it:
Indeed, all theorems are immediate corollaries of what we established (as
explained above); only the statements are longer, especially because I bother
expanding them.